requirements.txt
django==2.1.7
django-polymorphic==2.0.3
pkg-resources==0.0.0
pytz==2018.9
selenium==3.141.0
urllib3==1.24.1
